+++ Sources/StreamChatSwiftUI/Utils/Common/ChatClient+Extensions.swift
@@ -2,6 +2,7 @@
 // Copyright © 2024 Stream.io Inc. All rights reserved.
 //
 
+import Foundation
 import StreamChat
 
 extension ChatClient {