+++ Scripts/bootstrap.sh
@@ -40,8 +40,10 @@ chmod +x ./hooks/git-format-staged
 puts "Install brew dependencies"
 brew bundle -d
 
-puts "Bootstrap Mint dependencies"
-mint bootstrap --link
+if [ "${SKIP_MINT_BOOTSTRAP:-}" != true ]; then
+  puts "Bootstrap Mint dependencies"
+  mint bootstrap --link
+fi
 
 if [[ ${INSTALL_ALLURE-default} == true ]]; then
   puts "Install allurectl"