+++ Sources/StreamChatSwiftUI/ChatChannel/ChannelInfo/MediaAttachmentsView.swift
@@ -115,7 +115,7 @@ struct ImageAttachmentContentView: View {
             galleryShown = true
         } label: {
             LazyLoadingImage(
-                source: imageAttachment.imageURL,
+                source: MediaAttachment(url: imageAttachment.imageURL, type: .image),
                 width: itemWidth,
                 height: itemWidth
             )