+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/VideoAttachmentView.swift
@@ -161,11 +161,7 @@ struct VideoAttachmentContentView: View {
 
                 if width > 64 && attachment.uploadingState == nil {
                     VStack {
-                        Image(uiImage: images.playFilled)
-                            .customizable()
-                            .frame(width: 24)
-                            .foregroundColor(.white)
-                            .modifier(ShadowModifier())
+                        VideoPlayIcon()
                     }
                     .frame(width: width, height: width * ratio)
                     .contentShape(Rectangle())
@@ -202,5 +198,19 @@ struct VideoAttachmentContentView: View {
                 }
             }
         }
+    }
+}
+
+struct VideoPlayIcon: View {
+    @Injected(\.images) var images
+    
+    var width: CGFloat = 24
+    
+    var body: some View {
+        Image(uiImage: images.playFilled)
+            .customizable()
+            .frame(width: width)
+            .foregroundColor(.white)
+            .modifier(ShadowModifier())
     }
 }