+++ StreamChatSwiftUITestsAppTests/Robots/UserRobot.swift
@@ -29,33 +29,34 @@ final class UserRobot: Robot {
         ChannelListPage.userAvatar.safeTap()
         return self
     }
-
+    
     @discardableResult
-    func openChannel(channelCellIndex: Int = 0) -> Self {
-        let minExpectedCount = channelCellIndex + 1
-        let cells = ChannelListPage.cells.waitCount(minExpectedCount)
+    func waitForChannelListToLoad() -> Self {
+        let timeout = 15.0
+        let cells = ChannelListPage.cells.waitCount(1, timeout: timeout)
 
         // TODO: CIS-1737
         if !cells.firstMatch.exists {
             for _ in 0...10 {
-                server.stop()
                 app.terminate()
-                _ = server.start(port: in_port_t(MockServerConfiguration.port))
+                server.stop()
+                _ = server.start(port: UInt16(MockServerConfiguration.port))
                 sleep(1)
                 app.launch()
                 login()
-                cells.waitCount(minExpectedCount)
+                cells.waitCount(1, timeout: timeout)
                 if cells.firstMatch.exists { break }
             }
         }
 
-        XCTAssertGreaterThanOrEqual(
-            cells.count,
-            minExpectedCount,
-            "Channel cell is not found at index #\(channelCellIndex)"
-        )
-
-        cells.allElementsBoundByIndex[channelCellIndex].safeTap()
+        XCTAssertGreaterThanOrEqual(cells.count, 1, "Channel list has not been loaded")
+        return self
+    }
+    
+    @discardableResult
+    func openChannel(channelCellIndex: Int = 0) -> Self {
+        waitForChannelListToLoad()
+        ChannelListPage.cells.allElementsBoundByIndex[channelCellIndex].waitForHitPoint().safeTap()
         return self
     }
 }
@@ -362,6 +363,7 @@ extension UserRobot {
             sendMessage(text, waitForAppearance: false)
         }
         if send { tapOnSendGiphyButton() }
+        MessageListPage.Attributes.actionButtons().firstMatch.wait()
         return self
     }