+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### 🔄 Changed
 
 # [4.57.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.57.0)
-_June 06, 2024_
+_June 07, 2024_
 
 ### ✅ Added
 - Add support for creating and rendering polls [#495](https://github.com/GetStream/stream-chat-swiftui/pull/495)