+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.58.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.58.0)
+_June 27, 2024_
+
 ### ✅ Added
 - Thread replies shown in channel indicator [#518](https://github.com/GetStream/stream-chat-swiftui/pull/518)